requirements.txt
advent-of-code-data
black
ipython
python-dotenv